+++ fili-core/src/main/java/com/yahoo/bard/webservice/async/jobs/payloads/DefaultJobPayloadBuilder.java
@@ -2,7 +2,13 @@
 // Licensed under the terms of the Apache license. Please see LICENSE file distributed with this work for terms.
 package com.yahoo.bard.webservice.async.jobs.payloads;
 
-import com.yahoo.bard.webservice.async.jobs.jobrows.DefaultJobField;
+import static com.yahoo.bard.webservice.async.jobs.jobrows.DefaultJobField.DATE_CREATED;
+import static com.yahoo.bard.webservice.async.jobs.jobrows.DefaultJobField.DATE_UPDATED;
+import static com.yahoo.bard.webservice.async.jobs.jobrows.DefaultJobField.JOB_TICKET;
+import static com.yahoo.bard.webservice.async.jobs.jobrows.DefaultJobField.QUERY;
+import static com.yahoo.bard.webservice.async.jobs.jobrows.DefaultJobField.STATUS;
+import static com.yahoo.bard.webservice.async.jobs.jobrows.DefaultJobField.USER_ID;
+
 import com.yahoo.bard.webservice.async.jobs.jobrows.JobRow;
 import com.yahoo.bard.webservice.web.ErrorMessageFormat;
 import com.yahoo.bard.webservice.web.JobRequestFailedException;
@@ -31,16 +37,16 @@ public class DefaultJobPayloadBuilder implements JobPayloadBuilder {
         Map<String, String> job = new LinkedHashMap<>();
 
         String ticket = jobRow.getId();
-        Map<String, String> fieldValueMap = jobRow.getRowMap();
 
-        job.put("query", fieldValueMap.get(DefaultJobField.QUERY.getName()));
+        job.put(QUERY.getName(), jobRow.get(QUERY));
         job.put("results", JobPayloadBuilder.getResultsUrl(ticket, uriInfo));
         job.put("syncResults", JobPayloadBuilder.getSyncResultsUrl(ticket, uriInfo));
         job.put("self", JobPayloadBuilder.getSelfUrl(ticket, uriInfo));
-        job.put("status", fieldValueMap.get(DefaultJobField.STATUS.getName()));
-        job.put("jobTicket", fieldValueMap.get(DefaultJobField.JOB_TICKET.getName()));
-        job.put("dateCreated", fieldValueMap.get(DefaultJobField.DATE_CREATED.getName()));
-        job.put("userId", fieldValueMap.get(DefaultJobField.USER_ID.getName()));
+        job.put(STATUS.getName(), jobRow.get(STATUS));
+        job.put(JOB_TICKET.getName(), jobRow.get(JOB_TICKET));
+        job.put(DATE_CREATED.getName(), jobRow.get(DATE_CREATED));
+        job.put(DATE_UPDATED.getName(), jobRow.get(DATE_UPDATED));
+        job.put(USER_ID.getName(), jobRow.get(USER_ID));
 
         //throw exception if any of the JobFields are missing in the map
         if (job.containsValue(null)) {
@@ -50,7 +56,7 @@ public class DefaultJobPayloadBuilder implements JobPayloadBuilder {
                     .map(Map.Entry::getKey)
                     .collect(Collectors.toSet());
 
-            String msg = ErrorMessageFormat.JOB_MAPPING_FAILED.format(ticket, fieldValueMap, missingFields);
+            String msg = ErrorMessageFormat.JOB_MAPPING_FAILED.format(ticket, jobRow.getRowMap(), missingFields);
             LOG.error(msg);
             throw new JobRequestFailedException(msg);
         }