+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/async/AsyncResultsReadySpec.groovy
@@ -1,0 +1,153 @@
+package com.yahoo.bard.webservice.async
+
+import com.yahoo.bard.webservice.async.jobs.jobrows.DefaultJobStatus
+import com.yahoo.bard.webservice.util.GroovyTestUtils
+
+import spock.util.concurrent.PollingConditions
+
+/**
+ * Verifies that when the results of an asynchronous request are ready, that the job status is updated to "success" and
+ * the results are accessible through both the {@code results} and {@code syncResults} links.
+ */
+class AsyncResultsReadySpec extends AsyncFunctionalSpec {
+    /*
+        This executes the following requests, and expects the following responses. All caps are placeholders.
+             Send:
+                 http://localhost:9998/data/shapes/day?dateTime=2016-08-30%2F2016-08-31&metrics=height&asyncAfter=0
+             Receive:
+                 {
+                     "self": "http://localhost:9998/jobs/gregUUID",
+                     "userId": "greg",
+                     "ticket": "gregUUID",
+                     "results": "http://localhost:9998/jobs/gregUUID/results",
+                     "syncResults": "http://localhost:9998/jobs/gregUUID/results?asyncAfter=never",
+                     "dateCreated": "DATETIME",
+                     "dateUpdated": "SAME_DATETIME_AS_ABOVE",
+                     "status": "pending"
+                 }
+             Send:
+                 http://localhost:9998/jobs/gregUUID/results?asyncAfter=never
+             Receive:
+                {
+                     "rows" : [
+                          {
+                            "dateTime" : "2016-08-30 00:00:00.000",
+                            "height" : 100
+                          }
+                     ]
+                }
+            Send:
+                 http://localhost:9998/jobs/gregUUID/results
+            Receive:
+                {
+                     "rows" : [
+                          {
+                            "dateTime" : "2016-08-30 00:00:00.000",
+                            "height" : 100
+                          }
+                     ]
+                }
+            Send:
+                 http://localhost:9998/jobs/gregUUID
+            Receive:
+                {
+                     "self": "http://localhost:9998/jobs/gregUUID",
+                     "userId": "greg",
+                     "ticket": "gregUUID",
+                     "results": "http://localhost:9998/jobs/gregUUID/results",
+                     "syncResults": "http://localhost:9998/jobs/gregUUID/results?asyncAfter=never",
+                     "dateCreated": "DATETIME",
+                     "dateUpdated": "SAME_DATETIME_AS_ABOVE",
+                     "status": "success"
+                }
+    */
+
+    static final String QUERY =
+            "http://localhost:9998/data/shapes/day?dateTime=2016-08-30%2F2016-08-31&metrics=height&asyncAfter=0"
+
+    @Override
+    Map<String, Closure<String>> getResultsToTargetFunctions() {
+        [
+                data: { "data/shapes/day" },
+                //By querying the syncResults link first, we wait until the results are ready, thanks to the
+                //BroadcastChannel. Thus, we wait until the results are ready before verifying that the job status
+                //has been updated correctly.
+                syncResults: { AsyncTestUtils.extractTargetFromField(it.data.readEntity(String), "syncResults") },
+                results: { AsyncTestUtils.extractTargetFromField(it.data.readEntity(String), "results") },
+                jobs: { AsyncTestUtils.buildTicketLookup(it.data.readEntity(String)) }
+        ]
+    }
+
+    @Override
+    Map<String, Closure<Void>> getResultAssertions() {
+        [
+                data: {
+                    assert it.status == 202
+                    AsyncTestUtils.validateJobPayload(it.readEntity(String), QUERY, DefaultJobStatus.PENDING.name)
+                },
+                syncResults: {
+                    assert it.status == 200
+                    assert GroovyTestUtils.compareJson(it.readEntity(String), getExpectedApiResponse())
+                },
+                results: {
+                    assert it.status == 200
+                    assert GroovyTestUtils.compareJson(it.readEntity(String), getExpectedApiResponse())
+                },
+                jobs: { response ->
+                    PollingConditions pollingConditions = new PollingConditions()
+                    pollingConditions.eventually {
+                        assert response.status == 200
+                        AsyncTestUtils.validateJobPayload(
+                                response.readEntity(String),
+                                QUERY,
+                                DefaultJobStatus.SUCCESS.name
+                        )
+                    }
+                }
+        ]
+    }
+
+    @Override
+    Map<String, Closure<Map<String, List<String>>>> getQueryParameters() {
+        [
+                data: {[
+                        metrics: ["height"],
+                        asyncAfter: ["0"],
+                        dateTime: ["2016-08-30/2016-08-31"]
+                ]},
+                syncResults: {
+                    AsyncTestUtils.extractQueryParameters(
+                            new URI(AsyncTestUtils.getJobFieldValue(it.data.readEntity(String), "syncResults"))
+                    )
+                },
+                results: {[:]},
+                jobs: {[:]},
+        ]
+    }
+
+    @Override
+    Closure<String> getFakeDruidResponse() {
+        return {
+            """[
+                    {
+                        "version" : "v1",
+                        "timestamp": "2016-08-30",
+                        "result" : {
+                            "height" : 100
+                        }
+                    }
+            ]"""
+        }
+    }
+
+    String getExpectedApiResponse() {
+        """{
+              "rows" : [
+                  {
+                    "dateTime" : "2016-08-30 00:00:00.000",
+                    "height" : 100
+                  }
+              ]
+        }"""
+    }
+}